+++ views/index.ejs
@@ -3,24 +3,26 @@
     <div class="post-container">
         <h1>Welcome! Create your first post</h1> 
         <ul id="postsList">
-            <% posts.forEach(post=> { %>
+            <% posts.forEach(post => { %>
             <li>
                 <h2 class="post-title">
                     <%= post.title %>
                 </h2>
                 <small>
-                    <%= post.date %>
+                    <%= new Date(post.date).toLocaleString() %> 
                 </small>
                 <p>
                     <%= post.content %>
                 </p>
-                <small>By: <%= post.author %> </small>
+                <small>By: <%= post.author %></small>
                 <a class="edit-post" href="/edit/<%= post.id %>">Edit</a>
-                <a class="delete-post" href="/api/posts/delete/<%= post.id %>">Delete</a>
+                
+                <button class="delete-post" onclick="deletePost('<%= post.id %>')">Delete</button>
             </li>
             <hr>
-        <% }); %>
+            <% }); %>
         </ul>
     </div>
+
 </body>
 <%- include ("partials/footer.ejs") %>
