--- views/view.ejs
@@ -1,30 +1,0 @@
-<%- include("partials/header.ejs") %>
-<body>
-    <div class="post-container">
-        <ul id="postsList">
-            <% posts.forEach(post => { %>
-            <li>
-                <h2 class="post-title">
-                    <%= post.title %>
-                </h2>
-                <small>
-                    <%= post.date %>
-                </small>
-                <p>
-                    <%= post.content %>
-                </p>
-                <small>
-                    By: <%= pos.author %>
-                </small>
-                <a class="edit-post" href="/edit/<%= post.id %>">Edit</a>
-                <a class="delete-post" href="/api/posts/delete/<%= post.id %>">Delete</a>
-            </li>
-            <% }); %>
-
-        </ul>
-
-    </div>
-
-</body>
-
-<%- include("partials/footer.ejs") %>
